+++ srchybrid/PreferencesDlg.cpp
@@ -1,5 +1,5 @@
 //this file is part of eMule
-//Copyright (C)2002-2024 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / https://www.emule-project.net )
+//Copyright (C)2002-2026 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / https://www.emule-project.net )
 //
 //This program is free software; you can redistribute it and/or
 //modify it under the terms of the GNU General Public License
@@ -154,13 +154,13 @@ void CPreferencesDlg::Localize()
 	if (GetPageTreeControl()) {
 		static const UINT uids[15] =
 		{
-			IDS_PW_GENERAL, IDS_PW_DISPLAY, IDS_CONNECTION, IDS_PW_PROXY, IDS_PW_SERVER,
-			IDS_PW_DIR, IDS_PW_FILES, IDS_PW_EKDEV_OPTIONS, IDS_STATSSETUPINFO, IDS_IRC,
-			IDS_MESSAGESCOMMENTS, IDS_SECURITY, IDS_SCHEDULER, IDS_PW_WS, IDS_PW_TWEAK
+			IDS_PW_GENERAL, IDS_PW_DISPLAY, IDS_CONNECTION, IDS_PW_PROXY, IDS_PW_SERVER
+			, IDS_PW_DIR, IDS_PW_FILES, IDS_PW_EKDEV_OPTIONS, IDS_STATSSETUPINFO, IDS_IRC
+			, IDS_MESSAGESCOMMENTS, IDS_SECURITY, IDS_SCHEDULER, IDS_PW_WS, IDS_PW_TWEAK
 		};
 
 		int c;
-		for (c = 0; c < _countof(uids); ++c)
+		for (c = 0; c < (int)_countof(uids); ++c)
 			LocalizeItemText(c, uids[c]);
 #if defined(_DEBUG) || defined(USE_DEBUG_DEVICE)
 		GetPageTreeControl()->SetItemText(GetPageTreeItem(c), _T("Debug"));
